api/python/pyproject.toml
[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[project]
name = "quilt3"
version = "7.3.1"
description = "Quilt: where data comes together"
readme = "README.rst"
requires-python = ">=3.9"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
    { name = "quiltdata", email = "support@quilt.bio" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "platformdirs>=2",
    "boto3>=1.21.7",
    "jsonlines==1.2.0",
    "PyYAML>=5.1",
    "requests>=2.12.4",
    "tenacity>=5.1.1,!=8.4.0",
    "tqdm>=4.32",
    "requests_futures==1.0.0",
    "jsonschema>=3,<5",
    "pydantic>=2.0.0,<3.0.0",
    "awscrt>=0.31.0",
]

[project.optional-dependencies]
pyarrow = [
    "numpy>=1.14.0",
    "pandas>=0.19.2",
    "pyarrow>=0.14.1",
]
anndata = [
    "anndata>=0.8.0",
]
catalog = [
    "quilt3_local>=2,<3",
    "asgiproxy>=0.2,<0.3",
    "uvicorn>=0.15,<0.18",
    "aiobotocore[boto3]>=2",
    "lxml-html-clean>=0.4",
]

[project.urls]
Homepage = "https://quilt.bio/"
Documentation = "https://docs.quilt.bio/#quilt-python-sdk"
Repository = "https://github.com/quiltdata/quilt"
Issues = "https://github.com/quiltdata/quilt/issues"
Changelog = "https://docs.quilt.bio/quilt-python-sdk/more/changelog"

[project.scripts]
quilt3 = "quilt3.main:main"

[tool.uv.build-backend]
module-root = ""
module-name = "quilt3"

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
    "pytest-cov",
    "pytest-env",
    "pytest-subtests",
    "pytest~=8.4",
    "responses",
    "ruff>=0.12.11",
    "lxml-html-clean>=0.4",
    "quilt3[pyarrow,anndata]",
]

[tool.uv]
conflicts = [
    [
        { group = "dev" },
        { extra = "catalog" },
    ],
]

[tool.ruff]
# Extends from root configuration
extend = "../../ruff.toml"

[tool.poe.tasks]
_ensure-clean.cmd = "git diff --exit-code"

test.help = "Run test suite"
test.cmd = "pytest --disable-warnings"

test-verbose.help = "Run tests with verbose output"
test-verbose.cmd = "pytest -v"

test-cov.help = "Run tests with coverage report"
test-cov.cmd = "pytest --cov=. --cov-report=term-missing"

lint.help = "Run Ruff linter"
lint.cmd = "ruff check ."

fmt.help = "Format code with Ruff (append --check to check formatting)"
fmt.cmd = "ruff format"

clean.help = "Remove build artifacts and cache files"
clean.shell = "rm -rf build dist *.egg-info .pytest_cache .coverage* coverage.* && find . -type d -name __pycache__ -exec rm -rf {} + 2>/dev/null || true"

_gql-clean.shell = "rm -rf quilt3/_graphql_client"

_gql-gen.cmd = "uv run ariadne-codegen client"
_gql-gen.cwd = "quilt3-graphql"

_gql-fmt.cmd = "ruff format quilt3/_graphql_client"
_gql-lint-fix.cmd = "ruff check --fix quilt3/_graphql_client"

gql-gen.help = "Generate GraphQL client code for quilt3"
gql-gen.sequence = [
    "_gql-clean",
    "_gql-gen",
    "_gql-lint-fix",
    "_gql-fmt",
]

gql-check.help = "Check if GraphQL client code is up-to-date"
gql-check.sequence = [
    "gql-gen",
    "_ensure-clean",
]

verify-release-tag.help = "Verify git tag matches package version (for releases)"
verify-release-tag.args = [{ name = "tag", positional = true, required = true }]
verify-release-tag.shell = """
version=$(uv version --short)
if [ "$tag" != "$version" ]; then
  echo "Error: Git tag '$tag' does not match version '$version'"
  exit 1
fi
"""
verify-release-tag.verbosity = -1

gendocs.help = "Generate API documentation"
gendocs.cmd = "uv run --locked build.py"
gendocs.cwd = "../../gendocs"

gendocs-check.help = "Check if generated docs are up-to-date"
gendocs-check.sequence = [
    "gendocs",
    "_ensure-clean",
]

testdocs.help = "Test documentation code blocks"
testdocs.cmd = "uv run --locked pytest --codeblocks ../docs"
testdocs.cwd = "../../testdocs"

_catalog-test-config.help = "Write LOCAL frontend config for filesystem catalog testing"
_catalog-test-config.shell = """
mkdir -p ../../catalog/static-dev
cat > ../../catalog/static-dev/config.js <<'EOF'
window.QUILT_CATALOG_CONFIG = {
  region: 'us-east-1',
  registryUrl: '/__reg',
  alwaysRequiresAuth: false,
  apiGatewayEndpoint: '/__lambda',
  s3Proxy: '/__s3proxy',
  passwordAuth: 'DISABLED',
  ssoAuth: 'DISABLED',
  ssoProviders: '',
  stackVersion: 'local-dev',
  mode: 'LOCAL',
  mixpanelToken: '',
  serviceBucket: '',
  noDownload: false,
}
EOF
"""

_catalog-test-python.help = "Sync api/python with catalog dependencies"
_catalog-test-python.cmd = "uv sync --python 3.11 --no-dev --extra catalog"

_catalog-test-frontend.help = "Install catalog frontend dependencies"
_catalog-test-frontend.cmd = "npm install"
_catalog-test-frontend.cwd = "../../catalog"

_catalog-test-stage.help = "Reset LOCAL fixture bucket and stage preview fixtures"
_catalog-test-stage.shell = """
rm -rf /tmp/quilt-local-data
mkdir -p /tmp/quilt-local-data/demo-bucket
uv run --python 3.11 --no-dev --extra catalog python -m tests.preview_fixtures /tmp/quilt-local-data/demo-bucket
"""

_catalog-test-print.help = "Print next-step commands for LOCAL catalog testing"
_catalog-test-print.shell = """
printf '%s\n' \
  '' \
  'Prepared LOCAL catalog test state.' \
  'Frontend: cd ../../catalog && PORT=3001 npm start' \
  'Backend: cd . && PYTHONPATH=$PWD QUILT_LOCAL_OBJECT_BACKEND=filesystem QUILT_LOCAL_DATA_DIR=/tmp/quilt-local-data QUILT_CATALOG_URL=http://localhost:3001 uv run --python 3.11 --no-dev --extra catalog quilt3 catalog --host localhost --port 3000 --no-browser'
"""

catalog-test.help = "Prepare the LOCAL catalog filesystem test state"
catalog-test.sequence = [
    "_catalog-test-config",
    "_catalog-test-python",
    "_catalog-test-frontend",
    "_catalog-test-stage",
    "_catalog-test-print",
]
